separate post content section from post page
